# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NovaContent AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Campaign {
  id: number;
  userId: number;
  name: string;
  productName: string;
  description: string;
  category: string;
  audience: string;
  goal: string;
  tone: string;
  /** @nullable */
  keywords?: string | null;
  /** @nullable */
  cta?: string | null;
  platforms: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface CampaignInput {
  name: string;
  productName: string;
  description: string;
  category: string;
  audience: string;
  goal: string;
  tone: string;
  keywords?: string;
  cta?: string;
  platforms: string;
}

export interface CampaignUpdate {
  name?: string;
  productName?: string;
  description?: string;
  category?: string;
  audience?: string;
  goal?: string;
  tone?: string;
  keywords?: string;
  cta?: string;
  platforms?: string;
  status?: string;
}

export interface GeneratedContent {
  id: number;
  campaignId: number;
  contentType: string;
  content: string;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export interface GeneratedImage {
  id: number;
  campaignId: number;
  imageUrl: string;
  prompt: string;
  imageType: string;
  createdAt: string;
}

export interface CampaignDetail {
  id: number;
  userId: number;
  name: string;
  productName: string;
  description: string;
  category: string;
  audience: string;
  goal: string;
  tone: string;
  /** @nullable */
  keywords?: string | null;
  /** @nullable */
  cta?: string | null;
  platforms: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  content: GeneratedContent[];
  images: GeneratedImage[];
}

export interface GenerateRequest {
  campaignId: number;
}

export interface GenerateStarted {
  taskId: string;
  campaignId: number;
}

export interface TaskStatus {
  taskId: string;
  campaignId: number;
  status: string;
  progress: number;
  /** @nullable */
  currentStep?: string | null;
  completedSteps?: string[];
  /** @nullable */
  error?: string | null;
}

export interface RegenerateRequest {
  campaignId: number;
  contentType: string;
}

export interface WeeklyCount {
  week: string;
  count: number;
}

export interface ContentTypeCount {
  contentType: string;
  count: number;
}

export interface AnalyticsSummary {
  totalCampaigns: number;
  totalImages: number;
  totalBlogs: number;
  totalSocialPosts: number;
  campaignsPerWeek: WeeklyCount[];
  contentTypeBreakdown: ContentTypeCount[];
  recentCampaigns: Campaign[];
}