# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NovaContent AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Campaign } from './campaign';
import type { ContentTypeCount } from './contentTypeCount';
import type { WeeklyCount } from './weeklyCount';

export interface AnalyticsSummary {
  totalCampaigns: number;
  totalImages: number;
  totalBlogs: number;
  totalSocialPosts: number;
  campaignsPerWeek: WeeklyCount[];
  contentTypeBreakdown: ContentTypeCount[];
  recentCampaigns: Campaign[];
}