# lib/api-zod/src/generated/types/contentTypeCount.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NovaContent AI API
 * OpenAPI spec version: 0.1.0
 */

export interface ContentTypeCount {
  contentType: string;
  count: number;
}