# lib/api-zod/src/generated/types/generateStarted.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NovaContent AI API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateStarted {
  taskId: string;
  campaignId: number;
}