# lib/api-zod/src/generated/types/generatedContent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NovaContent AI API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedContent {
  id: number;
  campaignId: number;
  contentType: string;
  content: string;
  /** @nullable */
  metadata?: string | null;
  createdAt: Date;
}